requirements.txt
# Kalshi API
requests>=2.31.0
websockets>=12.0

# Web scraping
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0
feedparser>=6.0.0
trafilatura>=1.6.0

# Data & ML
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
torch>=2.1.0
transformers>=4.35.0
nltk>=3.8.0
spacy>=3.7.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# API server
fastapi>=0.104.0
uvicorn>=0.24.0
python-socketio>=5.10.0

# GUI
streamlit>=1.28.0
plotly>=5.18.0
streamlit-autorefresh>=1.0.1

# Scheduling & alerts
apscheduler>=3.10.0
python-dateutil>=2.8.0

# LLM integration
openai>=1.3.0
anthropic>=0.7.0

# Utils
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.5.0
aiohttp>=3.9.0
